Frontend beginnings

+++ dadjokes/views.py
@@ -1,18 +1,21 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth import authenticate, login, logout
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
-
-# Create your views here.
-
 def loginPage(request):
-
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -27,12 +30,32 @@ def loginPage(request):
             return redirect("home")
         else:
             messages.error(request,"Username or Password Doesnt exist")
-    context ={}
+  
+    context = {'page': page }
     return render(request, 'dadjokes/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    page = 'register' 
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during regitration')
+
+    context = {'page' : page, 'form' : form}
+    
+    return render(request, 'dadjokes/login_register.html',context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -46,29 +69,61 @@ def home(request):
     
     topics = Topic.objects.all()
     room_count = rooms.count()
-    
-    context = { 'rooms':rooms, 'topics':topics, 'room_count':room_count }
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains=q))
+
+
+    context = { 'rooms':rooms, 'topics':topics, 'room_count':room_count, 'room_messages': room_messages }
     return render(request,'dadjokes/home.html',context)
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
-    context = { 'room':room }        
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+    context = { 'room':room , 'room_messages' : room_messages, 'participants': participants}        
     return render(request,'dadjokes/room.html',context)
 
+
+
+def userProfile(request, pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    room_messages = user.message_set.all()
+    topics = Topic.objects.all()
+    context = {'user': user, 'rooms':rooms, 'topics':topics, 'room_messages':room_messages}
+    return render(request, "dadjokes/profile.html", context)
+
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
-            form.save()
+            room = form.save(commit=False)
+            room.host = request.user
+            room.save()
             return redirect('home')
             
     context = {'form':form}
     return render(request,'dadjokes/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -78,9 +133,27 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request,'dadjokes/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
     return render(request, 'dadjokes/delete.html' ,{'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse("You are not allowed")
+
+    if request.method == "POST":
+        message.delete()
+        return redirect("home")
+    return render(request, 'dadjokes/delete.html' ,{'obj':message})
+
